+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.3.0] - 2023-10-25
+### Added
+- Added word and character appearances
+
+### Changed
+- The plugin is totally refactored and uses a promise.
+- Fix for fragments as Appearance items
+- Another fix for hidden items in speaker view
+
+
 ## [1.2.1] - 2023-07-22
 ### Added
 - Fix for hidden items in print and speaker view